# backend/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RustPress CMS API",
    "description": "API completa del CMS RustPress — construido con Rust + Actix-Web",
    "version": "1.0.0",
    "contact": {
      "name": "RustPress",
      "url": "https://github.com/johandavid77/rustpress"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/shop/products": {
      "get": {
        "tags": [
          "shop"
        ],
        "summary": "Listar productos",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_price",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Productos"
          }
        }
      }
    },
    "/shop/products/slug/{slug}": {
      "get": {
        "tags": [
          "shop"
        ],
        "summary": "Producto por slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto"
          }
        }
      }
    },
...
    "/uptime": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Stats de uptime",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Uptime %"
          }
        }
      }
    }
  }
}